+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Fabric/FabricProvider.cs
@@ -21,11 +21,17 @@ public class FabricProvider : IProvider
     public IEnumerable<IProject> Projects => FabricProjectFactory.Projects;
    
     public async Task<IEnumerable<IVersion>> GetVersions(
-        string projectName,
         VersionOptions? options = null, 
         CancellationToken cancellationToken = default)
     {
-        return await FabricVersionFactory.GetVersion(projectName, options ?? new VersionOptions(), cancellationToken);
+        var versionOptions = options ?? new VersionOptions();
+        var versions = new List<IVersion>();
+
+        foreach (var project in Projects)
+        {
+            versions.AddRange(await project.GetVersions(versionOptions, cancellationToken));
+        }
+
+        return versions;
     }
-    
 }