+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Mohist/Model/MohistProject.cs
@@ -8,4 +8,12 @@ public record MohistProject(
     string Name,
     string Description,
     string Url,
-    byte[] Logo) : IProject;
+    byte[] Logo) : IProject
+{
+    public async Task<IEnumerable<IVersion>> GetVersions(
+        VersionOptions? options = null,
+        CancellationToken cancellationToken = default!)
+    {
+        return await MohistVersionFactory.GetVersion(Name, options ?? new VersionOptions(), cancellationToken);
+    }   
+}