+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Mohist/MohistProvider.cs
@@ -21,10 +21,17 @@ public class MohistProvider : IProvider
     public IEnumerable<IProject> Projects => MohistProjectFactory.Projects;
    
     public async Task<IEnumerable<IVersion>> GetVersions(
-        string projectName,
         VersionOptions? options = null, 
         CancellationToken cancellationToken = default)
     {
-        return await MohistVersionFactory.GetVersion(projectName, options ?? new VersionOptions(), cancellationToken);
-    }
+        var versionOptions = options ?? new VersionOptions();
+        var versions = new List<IVersion>();
+
+        foreach (var project in Projects)
+        {
+            versions.AddRange(await project.GetVersions(versionOptions, cancellationToken));
+        }
+
+        return versions;
+    } 
 }