+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Pocketmine/Model/PocketmineProject.cs
@@ -1,4 +1,5 @@
 ﻿using MinecraftJars.Core.Projects;
+using MinecraftJars.Core.Versions;
 
 namespace MinecraftJars.Plugin.Pocketmine.Model;
 
@@ -7,4 +8,12 @@ public record PocketmineProject(
     string Name,
     string Description,
     string Url,
-    byte[] Logo) : IProject;
+    byte[] Logo) : IProject
+{
+    public async Task<IEnumerable<IVersion>> GetVersions(
+        VersionOptions? options = null,
+        CancellationToken cancellationToken = default!)
+    {
+        return await PocketmineVersionFactory.GetVersion(Name, options ?? new VersionOptions(), cancellationToken);
+    }        
+}