+++ MinecraftJars.Plugin/MinecraftJars.Plugin.Pocketmine/PocketmineProvider.cs
@@ -22,12 +22,17 @@ public class PocketmineProvider : IProvider
     public IEnumerable<IProject> Projects => PocketmineProjectFactory.Projects;
 
     public async Task<IEnumerable<IVersion>> GetVersions(
-        string projectName,
         VersionOptions? options = null, 
         CancellationToken cancellationToken = default)
     {
-        return await PocketmineVersionFactory.GetVersion(projectName, options ?? new VersionOptions(), cancellationToken);
+        var versionOptions = options ?? new VersionOptions();
+        var versions = new List<IVersion>();
+
+        foreach (var project in Projects)
+        {
+            versions.AddRange(await project.GetVersions(versionOptions, cancellationToken));
+        }
+
+        return versions;
     }
-    
-    
 }